fix bug: not dumping a new blocks after fork
The problem was describe in psql_tools bug:
https://gitlab.syncad.com/hive/psql_tools/-/issues/16
After investigation it was realized that it is an issue in
sql_serializer code. During development there was a problem with
redundant calling by the processors function
'block_num_rendezvous_trigger::report_complete_thread_stage', the
problem was masked temporary by adding variable
m_already_commited_blocks and was forgotten to
remove.
Redundant event was cuased by triggering data processor with no data
when the plugin was closed or reindex was finished.