fix wrong traversing events with hive.app_next_block
The fixed problem may be responsible for HAF crashes
during fork reported by users. In some circumstances
repeated in the test added with this commit a context back to
already processed event what is a violation of a HAF rule and leads
to problems with clearing events queue and reversible data.