Get rid of EXCEPTION WHEN undefined_function in functional tests
The testing framework ignored errors related to calling undefined
functions. This is great, because tests don't need to provide empty
functions for all combination of {haf_admin,test_hived,alice,bob}_test_{given,when,error,then}.

This creates a possible issue if the test calls an undefined function
(because of a typo or lack of namespace for example). In this case test
is reported as passed where in fact it isn't. This makes it easy to
write a test that passes by accident.

Instead of ignoring undefined_function exception, predefine all
function combinations before loading test file. These functions are
defined to raise an error for *_error case and to be empty for other
cases. Test script can then replace required functions as needed.