Update HAF submodule with ci-base-image registry fix
Points to HAF fix/ci-base-image-version branch which corrects the comment
to show the image is in common-ci-configuration. The tag py3.14-1 was
already correct - the issue was HAfAH's BUILDER_IMAGE_PATH pointing to
the wrong registry (haf instead of common-ci-configuration).

Once HAF MR !758 merges, this commit will be in HAF develop.